requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
aiogram>=3.13.0
sqlalchemy[asyncio]>=2.0.0
aiosqlite>=0.19.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
jinja2>=3.1.0
python-multipart>=0.0.6
pydantic>=2.8.0
pydantic-settings>=2.3.0
httpx>=0.27.0

